Move `Case` into its own file
